DB and BLZ import

+++ my-awesome-guild/src/main/java/io/neferupito/myawesomeguild/data/repository/wow/RealmRepository.java
@@ -3,5 +3,5 @@ package io.neferupito.myawesomeguild.data.repository.wow;
 import io.neferupito.myawesomeguild.data.domain.wow.server.Realm;
 import org.springframework.data.repository.CrudRepository;
 
-public interface RealmRepository extends CrudRepository<Realm, Long> {
+public interface RealmRepository extends CrudRepository<Realm, Integer> {
 }
